Decouple client from hepdata_converter
I've changed the way the tests run, so they run locally but connect to the webservice running in docker. This will help us to test that the py2 and py3 versions of the server and client are compatible.

Removing the dependency on the converter has resulted in some test code being duplicated from hepdata_converter. I've logged a separate issue https://github.com/orgs/HEPData/projects/1#card-34125366 to deal with that later on.